requirements.txt
langgraph
langchain
langchain-google-genai
langchain-community
google-generativeai
python-dotenv
pydantic
uvicorn
fastapi
sqlalchemy>=1.4,<2.0.36
pandas
numpy
pytest
pytest-asyncio